+++ grunt/config/npm-command.js
@@ -1,0 +1,21 @@
+/**
+ * Generated header by Cozen for generator-cozen-angular project
+ * Generated file npm-command on WebStorm
+ *
+ * Created by: Geoffrey "C0ZEN" Testelin
+ * Date: 22/12/2017
+ * Time: 15:19
+ * Version: 1.0.0
+ */
+module.exports = {
+	test   : {
+		options: {
+			cmd: 'test'
+		}
+	},
+	publish: {
+		options: {
+			cmd: 'publish'
+		}
+	}
+};