+++ jestSetup.js
@@ -1,0 +1,18 @@
+/**
+ * Generated header by Cozen for generator-cozen-angular project
+ * Generated file jestSetup on WebStorm
+ *
+ * Created by: Geoffrey "C0ZEN" Testelin
+ * Date: 22/12/2017
+ * Time: 15:43
+ * Version: 1.0.0
+ */
+module.exports = function () {
+	return {
+		setup(w) {
+			w.testFramework.configure({
+				setupTestFrameworkScriptFile: '<rootDir>/jestSetup.js'
+			});
+		}
+	};
+};